implementation of miner with new template
